requirements.txt
azure-ai-inference
azure-ai-projects
azure-identity
azure-core
ipykernel
python-dotenv
jsonref
semantic_kernel[azure]
fastapi
uvicorn
requests
numpy
agent-framework